# public/swagger/v1/swagger.json
{
	"components": {
...
		"schemas": {
			"Link": {
				"properties": {
					"rel": {
						"type": "string"
					},
					"href": {
						"type": "string"
					}
				},
				"required": [
					"rel",
					"href"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Meta": {
				"properties": {
					"publisher": {
						"type": "string"
					},
					"licence": {
						"type": "string"
					},
					"documentation": {
						"type": "string"
					},
					"version": {
						"type": "string"
					},
					"comment": {
						"type": "string"
					},
					"hasFormat": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"limit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"publisher",
					"licence",
					"documentation",
					"version",
					"comment",
					"hasFormat",
					"limit"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Establishment": {
				"properties": {
					"fhrsid": {
						"type": "number",
						"format": "double"
					},
					"localAuthorityBusinessId": {
						"type": "string"
					},
					"businessName": {
						"type": "string"
					},
					"businessType": {
						"type": "string"
					},
					"ratingValue": {
						"type": "string"
					},
					"ratingDate": {
						"type": "string",
						"format": "date-time"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"fhrsid",
					"localAuthorityBusinessId",
					"businessName",
					"businessType",
					"ratingValue",
					"ratingDate",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IEstablishmentsDetailPaginationGetResponse": {
				"properties": {
					"establishments": {
						"items": {
							"$ref": "#/components/schemas/Establishment"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"establishments",
					"meta",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Scores": {
				"properties": {
					"hygiene": {
						"type": "number",
						"format": "double"
					},
					"structural": {
						"type": "number",
						"format": "double"
					},
					"confidenceInManagement": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"hygiene",
					"structural",
					"confidenceInManagement"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Geocode": {
				"properties": {
					"longitude": {
						"type": "string"
					},
					"latitude": {
						"type": "string"
					}
				},
				"required": [
					"longitude",
					"latitude"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EstablishmentListDetail": {
				"properties": {
					"fhrsid": {
						"type": "number",
						"format": "double"
					},
					"changesByServerID": {
						"type": "number",
						"format": "double"
					},
					"localAuthorityBusinessID": {
						"type": "string"
					},
					"businessName": {
						"type": "string"
					},
					"businessType": {
						"type": "string"
					},
					"businessTypeID": {
						"type": "number",
						"format": "double"
					},
					"addressLine1": {
						"type": "string"
					},
					"addressLine2": {
						"type": "string"
					},
					"addressLine3": {
						"type": "string"
					},
					"addressLine4": {
						"type": "string"
					},
					"postCode": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"ratingValue": {
						"type": "string"
					},
					"ratingKey": {
						"type": "string"
					},
					"ratingDate": {
						"type": "string",
						"format": "date-time"
					},
					"localAuthorityCode": {
						"type": "string"
					},
					"localAuthorityName": {
						"type": "string"
					},
					"localAuthorityWebSite": {
						"type": "string"
					},
					"localAuthorityEmailAddress": {
						"type": "string"
					},
					"scores": {
						"$ref": "#/components/schemas/Scores"
					},
					"schemeType": {
						"type": "string"
					},
					"geocode": {
						"$ref": "#/components/schemas/Geocode"
					},
					"rightToReply": {
						"type": "string"
					},
					"distance": {
						"type": "number",
						"format": "double"
					},
					"newRatingPending": {
						"type": "boolean"
					},
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"fhrsid",
					"changesByServerID",
					"localAuthorityBusinessID",
					"businessName",
					"businessType",
					"businessTypeID",
					"addressLine1",
					"addressLine2",
					"addressLine3",
					"addressLine4",
					"postCode",
					"phone",
					"ratingValue",
					"ratingKey",
					"ratingDate",
					"localAuthorityCode",
					"localAuthorityName",
					"localAuthorityWebSite",
					"localAuthorityEmailAddress",
					"scores",
					"schemeType",
					"geocode",
					"rightToReply",
					"distance",
					"newRatingPending",
					"meta",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IEstablishmentsListGetResponse": {
				"properties": {
					"establishments": {
						"items": {
							"$ref": "#/components/schemas/EstablishmentListDetail"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"establishments",
					"meta",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IEstablishmentsByIdGetResponse": {
				"properties": {
					"fhrsid": {
						"type": "number",
						"format": "double"
					},
					"changesByServerID": {
						"type": "number",
						"format": "double"
					},
					"localAuthorityBusinessID": {
						"type": "string"
					},
					"businessName": {
						"type": "string"
					},
					"businessType": {
						"type": "string"
					},
					"businessTypeID": {
						"type": "number",
						"format": "double"
					},
					"addressLine1": {
						"type": "string"
					},
					"addressLine2": {
						"type": "string"
					},
					"addressLine3": {
						"type": "string"
					},
					"addressLine4": {
						"type": "string"
					},
					"postCode": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"ratingValue": {
						"type": "string"
					},
					"ratingKey": {
						"type": "string"
					},
					"ratingDate": {
						"type": "string",
						"format": "date-time"
					},
					"localAuthorityCode": {
						"type": "string"
					},
					"localAuthorityName": {
						"type": "string"
					},
					"localAuthorityWebSite": {
						"type": "string"
					},
					"localAuthorityEmailAddress": {
						"type": "string"
					},
					"scores": {
						"$ref": "#/components/schemas/Scores"
					},
					"schemeType": {
						"type": "string"
					},
					"geocode": {
						"$ref": "#/components/schemas/Geocode"
					},
					"rightToReply": {
						"type": "string"
					},
					"distance": {
						"type": "number",
						"format": "double"
					},
					"newRatingPending": {
						"type": "boolean"
					},
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"fhrsid",
					"changesByServerID",
					"localAuthorityBusinessID",
					"businessName",
					"businessType",
					"businessTypeID",
					"addressLine1",
					"addressLine2",
					"addressLine3",
					"addressLine4",
					"postCode",
					"phone",
					"ratingValue",
					"ratingKey",
					"ratingDate",
					"localAuthorityCode",
					"localAuthorityName",
					"localAuthorityWebSite",
					"localAuthorityEmailAddress",
					"scores",
					"schemeType",
					"geocode",
					"rightToReply",
					"distance",
					"newRatingPending",
					"meta",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IEstablishmentsDetailGetResponse": {
				"properties": {
					"establishments": {
						"items": {
							"$ref": "#/components/schemas/EstablishmentListDetail"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"establishments",
					"meta",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_SCHEME_TYPE_KEYS": {
				"enum": [
					"FHRS",
					"FHIS"
				],
				"type": "string"
			},
			"_RATING_OPERATOR_KEYS": {
				"enum": [
					"Equal",
					"LessThanOrEqual",
					"GreaterThanOrEqual"
				],
				"type": "string"
			},
			"_SORT_OPTIONS_KEYS": {
				"enum": [
					"rating",
					"desc_rating",
					"alpha",
					"desc_alpha",
					"distance"
				],
				"type": "string"
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"name": "authorization",
				"in": "header",
				"bearerFormat": "JWT"
			},
			"apiKeyAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "X-API-KEY"
			}
		}
	},
	"info": {
		"title": "Food Standards Agency API",
		"version": "v1",
		"description": "The FSA Food Alerts API provides access to current and recent Food Alerts: Allergy Alerts (AA), Product Recall Information Notices (PRIN) and Food Alerts for Action (FAFA). It provides applications with the facility to list alerts matching some filter criterion, and to retrieve a description of an alert.",
		"contact": {
			"name": "Oraclo LTD",
			"email": "admin@orcalo.co.uk",
			"url": "https://orcalo.co.uk/"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/v1/establishments/{pageNumber}/{pageSize}": {
			"get": {
				"operationId": "GetEstablishmentsDetailPagination",
				"responses": {
					"200": {
						"description": "IEstablishmentsDetailPaginationGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IEstablishmentsDetailPaginationGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns a basic list of establishments (id/name), details can be retrieved by calling establishments/{id}, page parameters allow for page number and size specification.",
				"tags": [
					"Establishments"
				],
				"security": [],
				"parameters": [
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "pageSize",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/establishments/list/{id}": {
			"get": {
				"operationId": "GetEstablishmentsListDetail",
				"responses": {
					"200": {
						"description": "IEstablishmentsListGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IEstablishmentsListGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns a list of establishments, selected by Id.",
				"tags": [
					"Establishments"
				],
				"security": [],
				"parameters": [
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/establishments/{id}": {
			"get": {
				"operationId": "GetEstablishmentsDetailById",
				"responses": {
					"200": {
						"description": "IEstablishmentsByIdGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IEstablishmentsByIdGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns a details of a single establishment, selected by Id.",
				"tags": [
					"Establishments"
				],
				"security": [],
				"parameters": [
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/establishments": {
			"get": {
				"operationId": "GetEstablishmentsDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IEstablishmentsDetailGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns a collection of establishment details, based on provided search parameters. All search parameters are optional.",
				"tags": [
					"Establishments"
				],
				"security": [],
				"parameters": [
					{
						"description": "Filter the results based on the Scheme Type (FHRS or FHIS) of the establishment (See SchemeTypes for more details). Valid options are <b>'FHRS'</b> and </b>'FHIS'.",
						"in": "query",
						"name": "schemeTypeKey",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/_SCHEME_TYPE_KEYS"
						}
					},
					{
						"description": "Operator for increasing the flexibility of the rating search. Valid options are <b>'LessThanOrEqual', 'GreaterThanOrEqual' and 'Equal'</b>, the default for this value is 'Equal' (See RatingOperators for more details). Note: These operators only function on FHRS Ratings.",
						"in": "query",
						"name": "ratingOperatorKey",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/_RATING_OPERATOR_KEYS"
						}
					},
					{
						"description": "Set the options for sort field and direction of the results (See SortOptions for more details). Valid options are <b>'Relevance', 'rating', 'desc_rating', 'alpha', 'desc_alpha' and 'distance'</b>",
						"in": "query",
						"name": "sortOptionKey",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/_SORT_OPTIONS_KEYS"
						}
					},
					{
						"description": "The page number of the result set to return.",
						"in": "query",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Size of the page of results to return. (Note: Maximum value is system capped, so less results maybe returned than requested)",
						"in": "query",
						"name": "pageSize",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Filter the results based on the FHIS/FHRS Rating of the establishment (See Ratings for more details). Valid options for <b>FHRS</b> are <b>'0', '1', '2', '3', '4' and '5'</b> and for <b>FHIS<b> are <b>'Pass', 'ImprovementRequired', 'AwaitingPublication', 'AwatingInspection' and 'Exempt'</b>",
						"in": "query",
						"name": "ratingKey",
						"required": false,
						"schema": {}
					},
					{
						"description": "Terms to search for within the Business address and postcode of the establishments.",
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Terms to search for within the Business name of the establishments.",
						"in": "query",
						"name": "address",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The longitude of the centre point of a spatial query.",
						"in": "query",
						"name": "longitude",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The latitude of the centre point of a spatial query.",
						"in": "query",
						"name": "latitude",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The max distance limit in miles from the centre of the spatial query. Note: This value maybe capped by the system to limit the size of the result set",
						"in": "query",
						"name": "maxDistanceLimit",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter the results based on the Business Type of the establishment (See BusinessTypes for more details).",
						"in": "query",
						"name": "businessTypeId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter the results based on the Local Authority of the establishment (See Authorities for more details).",
						"in": "query",
						"name": "localAuthorityId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter the results based on the Country of the establishment (See Countries for more details).",
						"in": "query",
						"name": "countryId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}